[#6372] Fix failed test case
